+++ docs/docs.go
@@ -6,8 +6,7 @@ import (
 	"os"
 	"path"
 
-	// CHNAGE ME: Change this to your own provider GitHub
-	"github.com/cloudquery/cq-provider-template/resources"
+	"github.com/cloudquery/cq-provider-template/resources/provider"
 
 	"github.com/cloudquery/cq-provider-sdk/provider/docs"
 )
@@ -27,7 +26,7 @@ func main() {
 
 	}
 
-	if err = docs.GenerateDocs(resources.Provider(), outputPath, true); err != nil {
+	if err = docs.GenerateDocs(provider.Provider(), outputPath, true); err != nil {
 		fmt.Fprintf(os.Stderr, "Failed to generate docs: %s\n", err)
 	}
 }
