+++ resources/provider/provider.go
@@ -1,9 +1,10 @@
-package resources
+package provider
 
 import (
 	"embed"
 	// CHANGEME: change the following to your own package
 	"github.com/cloudquery/cq-provider-template/client"
+	"github.com/cloudquery/cq-provider-template/resources"
 
 	"github.com/cloudquery/cq-provider-sdk/provider"
 	"github.com/cloudquery/cq-provider-sdk/provider/schema"
@@ -23,7 +24,7 @@ func Provider() *provider.Provider {
 		Configure: client.Configure,
 		ResourceMap: map[string]*schema.Table{
 			// CHANGEME: place here all supported resources
-			"demo_resource": DemoResource(),
+			"demo_resource": resources.DemoResource(),
 		},
 		Migrations: providerMigrations,
 		Config: func() provider.Config {
