+++ resources/provider/provider_test.go
@@ -1,0 +1,13 @@
+package provider_test
+
+import (
+	"testing"
+
+	"github.com/cloudquery/cq-provider-template/resources/provider"
+
+	"github.com/cloudquery/cq-provider-sdk/migration"
+)
+
+func TestMigrations(t *testing.T) {
+	migration.RunMigrationsTest(t, provider.Provider(), []string{"latest"})
+}
